Make creation file names consistent
