+++ azure-pipelines/setup.yml
@@ -6,15 +6,15 @@ pool:
   vmImage: ubuntu-20.04
 
 variables:
-- group: vars
+  - group: vars
 
 steps:
-- task: AzureCLI@2
-  inputs:
-    azureSubscription: 'ARMSC'
-    scriptType: 'bash'
-    scriptLocation: 'inlineScript'
-    inlineScript: |
+  - task: AzureCLI@2
+    inputs:
+      azureSubscription: "ARMSC"
+      scriptType: "bash"
+      scriptLocation: "inlineScript"
+      inlineScript: |
         echo '--- Setting variables ---'
         subscription_name="$(subscription_name)"
         subscription_id=$(az account list --query "[?name=='$subscription_name'].id | [0]" | jq . -r)
@@ -40,7 +40,7 @@ steps:
         az config set extension.use_dynamic_install=yes_without_prompt
 
         echo '--- Creating a synthetic dataset ---'
-        python src/auto_setup/synthetic_dataset_creation.py
+        python src/auto_setup/create_synthetic_dataset.py
 
         echo '--- Creating AML config ---'
         python src/auto_setup/write_aml_template_yaml.py --subscription $subscription_id --resource_group $resource_group --workspace $workspace_name --acr_name $acr_name
@@ -53,7 +53,7 @@ steps:
         az ml compute create --min-instances 0 --max-instances 2 --name gpu-1 --size Standard_NC6 --workspace-name $workspace_name --resource-group $resource_group --no-wait --type AmlCompute --idle-time-before-scale-down 1800
         az ml compute update --name gpu-1 --resource-group $resource_group --workspace-name $workspace_name --min-instances 1
         az ml compute update --name gpu-1 --resource-group $resource_group --workspace-name $workspace_name --min-instances 0
-        
+
         echo '--- Fetching AML storage account name ---'
         aml_storage_account_name_str=$(az ml datastore list --resource-group $resource_group --workspace-name $workspace_name --query "[0].{account_name:account_name}.account_name")
         aml_storage_account_name=$(echo $aml_storage_account_name_str | tr -d \")
@@ -67,14 +67,14 @@ steps:
 
         echo '--- Setting devops configuration ---'
         az devops configure --defaults organization=https://dev.azure.com/$(organization_name) project=$(project_name)
-        
+
         echo '--- Creating Custom Vision resource ---'
         res=$(az cognitiveservices list --resource-group $resource_group --query "[?name=='$cv_name'].name | [0]" | jq . -r)
         if ((${#res}>0)); then   echo "Custom Vision resource already exists! Skipping step."; else   az cognitiveservices account create --name $cv_name --resource-group $resource_group --kind CustomVision.Training --sku F0 --location westeurope; fi
-        
+
         echo '--- Getting Custom Vision key ---'
         key=$(az cognitiveservices account keys list --name $cv_name --resource-group $resource_group | jq -r ".key1")
-        
+
         echo '--- Creating keyvault ---'
         az config set defaults.location=westeurope
         res=$(az keyvault list --resource-group $resource_group --query "[?name=='$keyvault_name'].name | [0]" | jq . -r)
