+++ src/auto_setup/tests/test_create_synthetic_dataset.py
@@ -4,11 +4,11 @@ import tempfile
 import pandas as pd
 from math import ceil
 
-from synthetic_dataset_creation import make_directories, generate_dataset
+from create_synthetic_dataset import make_directories, generate_dataset
 from util import get_lastest_iteration
 
 
-class TestSyntheticDatasetCreation(unittest.TestCase):
+class TestCreateSyntheticDataset(unittest.TestCase):
     def setUp(self):
         self.test_dir = tempfile.TemporaryFile()
